refactpr: extract utils functions
